updated server.py
server py now sends to google drive and creates two folders:

-uploads
-preprocessed uploads

one is for the content the user first uploads then the second folder is for the generated preprocessed upload.

currently the uploaded file is saved, but the second folder is empty because i still need to implement langchain and pinecone.