update server.py
we've included some document query code to fetch answers from similar docs on pinecone based on user input. the code hasnt been fully tested yet.

we'll need to update the flow for the front end as well. currently, we log "upload successful" on console. to signal success for the user, we're going to take the user to a more traditional chat interface to signal that  they can not interact with their document embeddings.